Renaming D44 rule/test files
