+++ dataactvalidator/config/sqlrules/d44_detached_award_financial_assistance_1.sql
@@ -1,7 +1,6 @@
 -- AwardeeOrRecipientUniqueIdentifier is required for AssistanceType of 02, 03, 04, or 05 whose
 -- ActionDate after October 1, 2010.
--- When ActionType = A, must be active in the System for Award Management (SAM) on the ActionDate of the award.
--- When ActionType = B, C, or D, DUNS should exist in SAM but need not be active on the ActionDate (i.e., it may be expired).
+-- Must be a valid 9-digit DUNS number
 
 CREATE OR REPLACE function pg_temp.is_date(str text) returns boolean AS $$
 BEGIN
@@ -12,35 +11,20 @@ EXCEPTION WHEN others THEN
 END;
 $$ LANGUAGE plpgsql;
 
-WITH detached_award_financial_assistance_d44_2_{0} AS
-    (SELECT submission_id,
-        row_number,
-        assistance_type,
-        action_date,
-        action_type,
-        awardee_or_recipient_uniqu
-    FROM detached_award_financial_assistance
-WHERE submission_id = {0})
-
 SELECT
     row_number,
     assistance_type,
     action_date,
-    action_type,
     awardee_or_recipient_uniqu
-FROM detached_award_financial_assistance_d44_2_{0} AS dafa
-WHERE COALESCE(dafa.assistance_type, '') IN ('02', '03', '04', '05')
+FROM detached_award_financial_assistance AS dafa
+WHERE submission_id = {0}
+    AND COALESCE(assistance_type, '') IN ('02', '03', '04', '05')
     AND (CASE
-        WHEN pg_temp.is_date(COALESCE(dafa.action_date, '0'))
+        WHEN pg_temp.is_date(COALESCE(action_date, '0'))
         THEN
-            CAST(dafa.action_date as DATE)
+            CAST(action_date as DATE)
     END) > CAST('10/01/2010' as DATE)
-    AND dafa.action_type = 'A'
-    AND dafa.row_number NOT IN (
-        SELECT DISTINCT sub_dafa.row_number
-        FROM detached_award_financial_assistance_d44_2_{0} as sub_dafa
-            JOIN executive_compensation AS exec_comp
-            ON (CAST(sub_dafa.awardee_or_recipient_uniqu as float) IS NOT DISTINCT FROM CAST(exec_comp.awardee_or_recipient_uniqu as float)
-            AND CAST(sub_dafa.action_date as Date) >= exec_comp.activation_date
-            )
+    AND COALESCE(dafa.awardee_or_recipient_uniqu, '') NOT IN (
+        SELECT exec_comp.awardee_or_recipient_uniqu
+        FROM executive_compensation as exec_comp
     )
